Move RxImmediateSchedulerRule.kt into the viewmodel package.
